[infra] restructures folders in python-package
