+++ python-package/basedosdados/__init__.py
@@ -1,7 +1,7 @@
-from basedosdados.dataset import Dataset
-from basedosdados.storage import Storage
-from basedosdados.table import Table
-from basedosdados.download import (
+from basedosdados.upload.dataset import Dataset
+from basedosdados.upload.storage import Storage
+from basedosdados.upload.table import Table
+from basedosdados.download.download import (
     read_sql,
     download,
     read_table,
