+++ python-package/basedosdados/upload/dataset.py
@@ -3,7 +3,7 @@ from google.cloud import bigquery
 
 from google.api_core.exceptions import Conflict
 
-from basedosdados.base import Base
+from basedosdados.upload.base import Base
 
 
 class Dataset(Base):
