+++ python-package/basedosdados/validdation/exceptions.py
